move NDArrayScan helpers and unit tests into CoreAIShared. Rename SpeechParityTest to SpeechParity to remove confusion
